Added rawtext factory
